requirement.txt
Flask
pytesseract
python-docx
pillow
gunicorn